+++ src/containers/email-confirm-modal/EmailConfirmModal.js
