+++ src/containers/email-confirm-modal/email-confirmation-modal.style.js
